move content harvest, indexing, to shared_tasks
