+++ dags/ecs_content_harvester.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from airflow.decorators import dag
 from airflow.models.param import Param
 
-from rikolti.dags.shared_content_harvester import ContentHarvestEcsOperator
+from dags.shared_tasks.content_harvest_operators import ContentHarvestEcsOperator
 
 
 @dag(
